Issue didRequestPermission and Location request
Set requestedLocation to false is creating issue here.
In fact didRequestPermission is called only when the alert for asking
permission is displayed.
By removing this line, the delegate is called too when the user chose
to accept or refuse the location service.

I don’t know if requestedLocation becomes useless or not with this
update, so I prefer to not remove it.